Move `GlotpressDownaloder` in dedicated file
